requirements.txt
flask
flask-cors
google-api-python-client
google-auth-oauthlib
google-auth
openai